==> 21-Metabolome_DimensionReduction.Rmd <==
```{r, include = FALSE}
knitr::opts_chunk$set(out.width = "100%", message = FALSE, warning = FALSE)
```


# Dimension Reduction {#DimensionReduction}


降维分析是为了查看样本的差异或挖掘潜在标志物。方法有以下三种：

+ 主成分分析（PCA）

主成分分析PCA（Principal Component Analysis）是一种常用的数据分析方法。PCA通过线性变换将原始数据变换为一组各维度线性无关的表示，可用于提取数据的主要特征分量，常用于高维数据的降维。

+ 偏最小二乘回归分析法（PLS-DA）

偏最小二乘判别分析（PLS-DA）是一种有监督模式识别的多元统计分析方法，将多维数据在压缩前先按需要寻找的差异因素分组。

+ 正交偏最小二乘判别分析（OPLS-DA）

正交偏最小二乘判别分析（OPLS-DA）结合了正交信号矫正（OSC）和PLS-DA方法，能够将X矩阵信息分解成与Y相关和不相关的两类信息，通过去除不相关的差异来筛选差异变量。


## 加载R包
```{r, message = FALSE, warning = FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)
library(tidyverse)
library(ropls)
library(SummarizedExperiment)
library(MicrobiomeAnalysis)

# rm(list = ls())
options(stringsAsFactors = F)
options(future.globals.maxSize = 1000 * 1024^2)

grp_names <- c("None", "Mild", "Moderate", "Severe")
grp_colors <- c("#7DD06F", "#844081", "#688EC1", "#C17E73")
```


## 导入数据

对数据[OmicsDataSet-Zeybel et al. - 2022.xlsx](https://github.com/HuaZou/DraftNotes/blob/main/InputData/Zeybel-2022/OmicsDataSet-Zeybel et al. - 2022.xlsx)处理后生成的，可参考数据预处理章节。

> ```R
> saveRDS(se_scale, "./InputData/result/QC/se_scale.RDS", compress = TRUE)
> ```

```{r}
data_meta <- readRDS("./InputData/result/QC/se_scale.RDS")
```


## 函数
```{r}
DR_fun <- function(
    x,
    group,
    group_names,
    group_colors,
    DRtype = c("PCA", "PLS", "OPLS"),
    occ_cutoff = 0.5) {
  
  # x = data_meta
  # group = "LiverFatClass"
  # group_names = grp_names
  # group_colors = grp_colors
  # DRtype = "PLS" # PCA
  # occ_cutoff = 0.5
  
  # dataseat
  metadata <- SummarizedExperiment::colData(x) %>%
      as.data.frame()
  profile <- SummarizedExperiment::assay(x) %>%
      as.data.frame()

  colnames(metadata)[which(colnames(metadata) == group)] <- "CompVar"
  phenotype <- metadata %>%
    dplyr::filter(CompVar %in% group_names) %>%
    dplyr::mutate(CompVar = as.character(CompVar)) %>%
    dplyr::mutate(CompVar = factor(CompVar, levels = group_names))
  
  sid <- intersect(rownames(phenotype), colnames(profile))
  phen <- phenotype[pmatch(sid, rownames(phenotype)), , ]
  prof <- profile %>%
    dplyr::select(all_of(sid))
  
  if (!all(colnames(prof) == rownames(phen))) {
    stop("Wrong Order")
  }
  
  trim_FeatureOrSample <- function(x, nRow, threshold) {
  
    df_occ <- apply(x, nRow, function(x) {
      length(x[c(which(!is.na(x) & x!=0))]) / length(x)
    }) %>%
      data.frame() %>% stats::setNames("Occ") %>%
      tibble::rownames_to_column("type")
    if(nRow == 1){
      rownames(df_occ) <- rownames(x)
    }else{
      rownames(df_occ) <- colnames(x)
    }
    df_KEEP <- apply(df_occ > threshold, 1, all) %>%
      data.frame() %>% stats::setNames("Status") %>%
      dplyr::filter(Status)
  
    res <- x %>%
      tibble::rownames_to_column("featureid") %>%
      dplyr::filter(featureid %in% rownames(df_KEEP)) %>%
      tibble::column_to_rownames("featureid")
    
    return(res)
  }

  prof_cln <- trim_FeatureOrSample(prof, 1, occ_cutoff)
  dataMatrix <- prof_cln %>% t() # row->sampleID; col->features
  sampleMetadata <- phen %>% # row->sampleID; col->features
    dplyr::mutate(CompVar = factor(CompVar, levels = group_names)) %>%
    dplyr::mutate(Color = factor(CompVar, 
                                  levels = group_names,
                                  labels = group_colors),
                  Color = as.character(Color))    
  
  if (DRtype == "PCA") {
    fit <- opls(dataMatrix)
    plot(fit,
         typeVc = "x-score",
         parAsColFcVn = sampleMetadata$CompVar,
         )
  } else if (DRtype == "PLS") {
    fit <- opls(dataMatrix, sampleMetadata$CompVar, predI = 2)
    plot(fit,
         typeVc = "x-score",
         parAsColFcVn = sampleMetadata$CompVar,
         )
  } else if (DRtype == "OPLS") {
    # only for binary classification
    fit <- opls(dataMatrix, sampleMetadata$CompVar, predI = 1, orthoI = NA)
    plot(fit,
         typeVc = "x-score",
         parAsColFcVn = sampleMetadata$CompVar,
         )
  }

  return(fit)
}
```


## 主成分分析（PCA）
```{r, fig.width=6, fig.height=5}
PCA_res <- DR_fun(
    x = data_meta,
    group = "LiverFatClass",
    group_names = grp_names,
    group_colors = grp_colors,    
    DRtype = "PCA",
    occ_cutoff = 0.5)
```


结果：PCA将原始features重新组合成新的主成分，进而形成了基于主成分的矩阵。每个主成分是由原来features的线性组合而成。

+ **左上图**: 碎石图展示不同主成分的可解释度，也即是代表全体的features，评估主成分是否足够

+ **右上图**: 离群点图，它展示的各样本在投影平面和投影平面正交的距离，数值较高样本表明它们与其他样本间的差异较大 

+ **左下图**：每个样本的主成分得分（可简单理解为系数，此处只展示第一和第二主成分），该数值表示主成分能解释多少features: 该数值由features矩阵分解得到。各个样本在PC1和PC2轴的排序坐标差异，通过它们可评估样本在组成上的差异。

+ **右下图**：载荷值: 理解为构成每个主成分的features，可以认为一个主成分是由N个feature线性组合合成，每个features对该主成分的贡献程度作为loading的数值。边缘处的变量表示它们在各样本中的含量差别明显（如在某些样本中具有较大/较小的极端值等），即对排序空间的贡献较大，暗示它们可能为一些重要的代谢物。



## 偏最小二乘回归分析法（PLS-DA）

PLS-DA需要提供Y响应变量，它通过投影分别将预测变量（Y响应变量）和观测变量（自变量）投影到一个新空间，来寻找一个线性回归模型。通过建立组学数据与样本类别之间的关系模型，实现对样本类别的预测，为有监督的建模方式。

偏最小二乘（PLS）是一种基于预测变量和响应变量之间协方差的潜在变量回归方法，找到不同类别的分割信息最优的线性组合，PLS的每一个成分都赋予一个权值，即对应该成分的分离能力。


```{r, fig.width=6, fig.height=5}
PLS_res <- DR_fun(
    x = data_meta,
    group = "LiverFatClass",
    group_names = grp_names,
    group_colors = grp_colors,    
    DRtype = "PLS",
    occ_cutoff = 0.5)
```

结果：R2X和R2Y分别表示所建模型对X和Y矩阵的解释率，Q2标示模型的预测能力，它们的值越接近于1表明模型的拟合度越好，训练集的样本越能够被准确划分到其原始归属中。

+ **左上图**: 展示了2个正交轴的R2Y和Q2Y

+ **右上图**: PLS-DA模型的R2Y和Q2Y与随机置换数据后获得的相应值进行比较 

+ **左下图**：展示了各样本在投影平面内以及正交投影面的距离，具有高值的样本标注出名称，表明它们与其它样本间的差异较大。颜色代表性别分组。

+ **右下图**：各样本在PLS-DA轴中的坐标，颜色代表不同分组。我们可以看到，相对于上文的PCA（仅通过方差特征值分解），PLS-DA在区分组间差异时更有效（带监督的偏最小二乘判别分析）。图的下方还提供了R2X、R2Y等值，用于评估模型优度。


此外，还可通过变量投影重要度（Variable Importance for the Projection，VIP）衡量各代谢物组分含量对样本分类判别的影响强度和解释能力，辅助标志代谢物的筛选。通常以**VIP > 1**作为筛选标准。

```{r}
vip_values <- getVipVn(PLS_res)

vip_select <- vip_values[vip_values > 1]

head(vip_select)
```



## 正交偏最小二乘判别分析（OPLS-DA）

PLS-DA容易出现过拟合（overfitting）问题。所谓过拟合，即通过训练集建立了一个预测模型，它在训练集上表现出色，但通过测试集测试时却表现不佳。过拟合是机器学习中的一个常见问题，主要出现在具有比样本数量更多的变量数量的数据集的分析中。主要是因为PLS-DA是有监督的预测模型且偏向训练数据。

相比PLS-DA，叠加了正交分解的OPLS-DA能更好地避免过拟合现象，这是因为它会通过将与X和Y正交的变量删除，避免不差异的变量影响结果。OPLS从给定的自变量数据集中移除正交变量，并把这些正交变量和非正交变量区分开来。然后再对非正交变量做偏最小二乘法获得载荷矩阵也即是在Y情况下的最大方差。


+ None vs Severe
```{r, fig.width=6, fig.height=5}
DR_fun(
    x = data_meta,
    group = "LiverFatClass",
    group_names = grp_names[c(1, 4)],
    group_colors = grp_colors[c(1, 4)],
    DRtype = "OPLS",
    occ_cutoff = 0.5)
```


结果和PLS-DA一致。

+ **左上图**: 展示了2个正交轴的R2Y和Q2Y

+ **右上图**: PLS-DA模型的R2Y和Q2Y与随机置换数据后获得的相应值进行比较 

+ **左下图**：展示了各样本在投影平面内以及正交投影面的距离，具有高值的样本标注出名称，表明它们与其它样本间的差异较大。颜色代表性别分组。

+ **右下图**：各样本在PLS-DA轴中的坐标，颜色代表不同分组。我们可以看到，相对于上文的PCA（仅通过方差特征值分解），PLS-DA在区分组间差异时更有效（带监督的偏最小二乘判别分析）。图的下方还提供了R2X、R2Y等值，用于评估模型优度。



## Session info
```{r}
devtools::session_info()
```


## Reference

+ [ropls: PCA, PLS(-DA) and OPLS(-DA) for multivariate analysis and feature selection of omics data](https://master.bioconductor.org/packages/release/bioc/vignettes/ropls/inst/doc/ropls-vignette.html)
